google logout and api security

+++ ProMa/src/com/au/proma/controller/BuController.java
@@ -2,6 +2,7 @@ package com.au.proma.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,40 +26,65 @@ public class BuController {
 	
 	@RequestMapping(method=RequestMethod.GET, produces="application/json")
 	@ResponseBody
-	public List<BU> getAllBU(){
-		return buService.getAllBU();
+	public List<BU> getAllBU(HttpServletRequest request){
+		if(request.getSession().getAttribute("set").equals("true")==true)
+		{
+			return buService.getAllBU();
+		}
+		else
+		{
+			return null;
+		}
+		
 	}
 	
 	@RequestMapping(method=RequestMethod.POST, produces="application/json",consumes="application/json")
 	@ResponseBody
-	public Integer addBU(@RequestBody BU bu){
+	public Integer addBU(@RequestBody BU bu,HttpServletRequest request){
+		if(request.getSession()!=null&&request.getSession().getAttribute("role").equals("admin")==true)
 		return buService.addBU(bu);
+		else
+			return null;
 	}
 	
 	@RequestMapping(value="/{buid}",method=RequestMethod.DELETE, produces="application/json")
 	@ResponseBody
-	public Boolean removeBU(@PathVariable("buid") int buid){
-		BU bu = new BU();
+	public Boolean removeBU(@PathVariable("buid") int buid,HttpServletRequest request){
+		if(request.getSession()!=null&&request.getSession().getAttribute("role").equals("admin")==true)
+		{BU bu = new BU();
 		bu.setBuid(buid);
-		return buService.removeBU(bu);
+		return buService.removeBU(bu);}
+		else
+			return null;
 	}
 	
 	@RequestMapping(value="/{buid}/buheads",method=RequestMethod.POST, produces="application/json",consumes="application/json")
 	@ResponseBody
-	public Boolean addBUHead(@RequestBody User user,@PathVariable("buid") int buid){
-		BU bu = new BU();
-		bu.setBuid(buid);
-		return buService.addBUHead(bu, user);
+	public Boolean addBUHead(@RequestBody User user,@PathVariable("buid") int buid,HttpServletRequest request){
+		if(request.getSession()!=null&&request.getSession().getAttribute("role").equals("admin")==true)
+		{
+			BU bu = new BU();
+			bu.setBuid(buid);
+			return buService.addBUHead(bu, user);
+		}
+		else
+			return null;
 	}
 
 	@RequestMapping(value="/{buid}/buheads/{userid}",method=RequestMethod.DELETE, produces="application/json",consumes="application/json")
 	@ResponseBody
-	public Boolean addBUHead(@PathVariable("buid") int buid,@PathVariable("userid") int userid){
-		BU bu = new BU();
-		bu.setBuid(buid);
-		User user = new User();
-		user.setUserid(userid);
-		return buService.removeBUHead(bu, user);
+	public Boolean addBUHead(@PathVariable("buid") int buid,@PathVariable("userid") int userid,HttpServletRequest request){
+		if(request.getSession()!=null&&request.getSession().getAttribute("role").equals("admin")==true)
+		{		
+			BU bu = new BU();
+			bu.setBuid(buid);
+			User user = new User();
+			user.setUserid(userid);
+			return buService.removeBUHead(bu, user);
+	
+		}
+		else 
+			return null;
 	}
